handle create order and orderItem at local, improve ui Status when losing connection

+++ smartPOS/Common/Models/OrderItemTopping.swift
@@ -13,5 +13,5 @@ struct OrderItemTopping: Decodable {
     let name: String?
     let menuItemToppingId: String?
     let quantity: Int?
-    let price: Float?
+    let price: Double?
 }
